dres0280: change RAM limit to 384MiByte
We do not use GPU on board dres0280 and
only 256MiByte RAM are usable on dres0280_rev_a.

RAM limit is currently set to 384MiByte.
This is a hack that maps the GPU memory to the area starting
at 256MiByte. This ensures that all of our RAM is usable by
the rest of the system.

WARNING: Do never access the GPU memory as there is no usable
memory assigned to it.

Signed-off-by: Eik Binschek <binschek@dresearch-fe.de>